Update Constraint Record to match the server response.
